# [LinearSolver.Direct] Better distribution of tasks among threads