I worked with the scene `examples\Demos\fallingBeamLagrangianCollision.scn`


I noticed that the distribution of the work among threads is not uniform:

![image](https://github.com/sofa-framework/sofa/assets/10572752/9145a080-0a7a-4b22-b3dc-501d509aefcf)


Here is the distribution after the changes:

![image](https://github.com/sofa-framework/sofa/assets/10572752/cd047f43-aa7c-49de-a303-7d1945d8dcbd)


Benchmark on 500 time steps:
Before:
```
[INFO]    [BatchGUI] 500 iterations done in 22.9377 s ( 21.7982 FPS).
 LEVEL   START    NUM      MIN     MAX   MEAN     DEV    TOTAL  PERCENT ID
   5      14.42    1       0.01   30.28   19.39    6.24   19.39   42.69 .....Get Compliance
```

After
```
[INFO]    [BatchGUI] 500 iterations done in 19.7138 s ( 25.363 FPS).
 LEVEL   START    NUM      MIN     MAX   MEAN     DEV    TOTAL  PERCENT ID
  5      13.77    1       0      22.02   15.31    5.07   15.31   39.29 .....Get Compliance
```


______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).