@bakpaul 2023-10-10
On a quite heavy scene of mine with five SparseLDLSolvers, I have the following results : 

Before:
```
[INFO]    [BatchGUI] 1000 iterations done in 118.746 s ( 8.42137 FPS).                                                                                                                                            
 LEVEL	 START	  NUM	   MIN	   MAX	 MEAN	  DEV	 TOTAL	PERCENT	ID
   5   	  25.65	   1   	  90.97	 136.87	  97.36	   9.53	  97.36	  76.30	.....Get Compliance
```

After:
```
[INFO]    [BatchGUI] 1000 iterations done in 115.418 s ( 8.6642 FPS). 
 LEVEL	 START	  NUM	   MIN	   MAX	 MEAN	  DEV	 TOTAL	PERCENT	ID
   5   	  25   	   1   	  88.55	 104.93	  93.89	   3.04	  93.89	  75.91	.....Get Compliance
```

So there's an improvement both in the percentage and in the standard deviation meaning that it is more stable and thus the distribution seems to be more efficient. Nice !