@alxbilger 2023-10-10
@bakpaul So much work for only 3 FPS 😭